refactor: 설정 파일 패키지 변경
