+++ src/main/java/chilling/global/config/swagger/SwaggerConfig.java
@@ -1,4 +1,4 @@
-package chilling.encore.config.swagger;
+package chilling.global.config.swagger;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
